Removed testify dependency from util file
