+++ go/store/chunks/test_utils.go
@@ -25,24 +25,9 @@ import (
 	"context"
 	"sync/atomic"
 
-	"github.com/stretchr/testify/assert"
-
 	"github.com/dolthub/dolt/go/store/d"
 	"github.com/dolthub/dolt/go/store/hash"
 )
-
-func assertInputInStore(input string, h hash.Hash, s ChunkStore, assert *assert.Assertions) {
-	chunk, err := s.Get(context.Background(), h)
-	assert.NoError(err)
-	assert.False(chunk.IsEmpty(), "Shouldn't get empty chunk for %s", h.String())
-	assert.Equal(input, string(chunk.Data()))
-}
-
-func assertInputNotInStore(input string, h hash.Hash, s ChunkStore, assert *assert.Assertions) {
-	chunk, err := s.Get(context.Background(), h)
-	assert.NoError(err)
-	assert.True(chunk.IsEmpty(), "Shouldn't get non-empty chunk for %s: %v", h.String(), chunk)
-}
 
 type TestStorage struct {
 	MemoryStorage
